+++ dev/build.image/notes.txt
@@ -1,0 +1,42 @@
+open liberty: open-liberty_image_fat
+WS-CD: liberty_image_fat
+
+gradlew releaseNeeded
+
+gradle startup properties need to have these:
+
+c:/Users/ThomasBitonti/gradle.startup.properties
+
+#buildResultUUID=XXXXJ
+#buildLabel=ChucksBuild
+#is.automated.build=true
+
+[   ]    openliberty-MavenArtifact-22.0.0.3-202202142326.zip    2022-02-15 10:35    451M
+"runtime" TEST
+
+installation is extraction; not running the jar
+the resulting image should be fully runnable
+
+(1) extract/create/run as initial tests
+
+[   ]    openliberty-22.0.0.3-202202142326.zip    2022-02-15 10:35    275M
+[   ]    openliberty-22.0.0.4-beta-202202142326.zip    2022-02-15 10:35    283M // includes beta
+TEST
+
+[   ]    openliberty-all-22.0.0.3-202202142326.zip    2022-02-15 10:36    312M  // includes beta + no_ship
+[   ]    openliberty-dev-22.0.0.3-202202142326.zip    2022-02-15 10:36    320M  // includes beta + no_ship + test infra
+
+(2) verify profile feature sets
+
+(3) do feature install starting with the kernel
+
+TEST
+[   ]    openliberty-jakartaee9-22.0.0.3-202202142326.zip    2022-02-15 10:36    104M
+[   ]    openliberty-javaee8-22.0.0.3-202202142326.zip    2022-02-15 10:36    119M
+[   ]    openliberty-kernel-22.0.0.3-202202142326.zip    2022-02-15 10:36    11M
+[   ]    openliberty-microProfile4-22.0.0.3-202202142326.zip    2022-02-15 10:36    51M
+[   ]    openliberty-microProfile5-22.0.0.3-202202142326.zip    2022-02-15 10:36    47M
+[   ]    openliberty-webProfile8-22.0.0.3-202202142326.zip    2022-02-15 10:36    89M
+[   ]    openliberty-webProfile9-22.0.0.3-202202142326.zip    2022-02-15 10:36    84M
+
+look in build.image/build/libs/...
