+++ dev/build.image_fat/.classpath
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
+	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="src" path="fat"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/fattest.simplicity"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/wlp.lib.extract"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/com.ibm.ws.install"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/junit"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
