+++ dev/build.image_fat/fat/com/ibm/ws/image/test/util/ScriptFilter.java
