+++ dev/build.image_test_ol/.classpath
@@ -7,5 +7,6 @@
 	<classpathentry combineaccessrules="false" kind="src" path="/com.ibm.ws.install"/>
 	<classpathentry combineaccessrules="false" kind="src" path="/wlp.lib.extract"/>
 	<classpathentry kind="lib" path="C:/dev/repos-pub/o-l/dev/fattest.simplicity/build/libs/lib/hamcrest-all-1.3.jar"/>
+	<classpathentry combineaccessrules="false" kind="src" path="/build.image_fat"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
